Throw an exception on reading a non-wiki page with wiki api
Wiki pages are nodes.
Nodes have unique key, usually an MD5 hash of something.
Wiki pages have it hashing the page title.
This allows for easy page picking by name.
If a node is read by the key and is suddenly not a wiki -- fail.

TODO: use an index, not the key.